prevent potential deadlock
Because there is a single condition variable in the ArrayPool, when
Get() calls Signal() to unlock a single go routine waiting on the
condition variable, that particular go routine that is unlocked might
be another Get().

Same thing when Put() calls Signal() to unlock, it might just unlock
another go routine waiting on Put().

Changing the call from Signal() to Broadcast() means that all of the
go routines blocked on that condition variable are unlocked, which is
less efficient, however it will not result in deadlock.