added SemaphorePool to test code with dual condition variables
Because ArrayPool invokes the Broadcast() method on its single
condition variable, and unlocks all go routines waiting on the
condition variable, it unlocks all go routines waiting on Get() and
all go routines waiting on Put(). This is not ideal, as after Get()
has run, we know it only makes sense to unlock Put(), and vice versa.

This code adds a new structure called SemaphorePool to the test code,
and it seems to work, but was not integrated because it has a slightly
different API.

The SemaphorePool has two condition variables, and Get() will unlock
only the go routines blocked on Put(), and vice versa. Note that both
condition variables use the same underlying mutex because they both
need to run exclusively of any other as they work with a single free
list.